Extract waitInitMediator to NewMessageUtils so it lightens a bit NewMessageFragment
